add AdamOptimizer, fix datatype

+++ elasticdl/pkg/kernel/capi/kernel_api.h
@@ -11,12 +11,12 @@ void Adam(float* grad,
           float* param,
           float* m,
           float* v,
-          double lr,
+          float lr,
           long long size,
           long long step,
-          double beta1,
-          double beta2,
-          double epsilon,
+          float beta1,
+          float beta2,
+          float epsilon,
           float* max_square);
 
 #ifdef __cplusplus
